Move SymbolTable to mypyc_hacks
