+++ mypy/mypyc_hacks.py
@@ -35,3 +35,59 @@ class TypeOfAny(Enum):
     from_another_any = 'from_another_any'
     # Does this Any come from an implementation limitation/bug?
     implementation_artifact = 'implementation_artifact'
+
+
+# Moved from nodes.py, because it inherits from dict
+from typing import Dict, List, Any
+
+JsonDict = Dict[str, Any]
+
+MYPY = False
+if MYPY:
+    from mypy.nodes import SymbolTableNode
+
+class SymbolTable(Dict[str, 'SymbolTableNode']):
+    def __str__(self) -> str:
+        from mypy.nodes import implicit_module_attrs, SymbolTableNode
+
+        a = []  # type: List[str]
+        for key, value in self.items():
+            # Filter out the implicit import of builtins.
+            if isinstance(value, SymbolTableNode):
+                if (value.fullname != 'builtins' and
+                        (value.fullname or '').split('.')[-1] not in
+                        implicit_module_attrs):
+                    a.append('  ' + str(key) + ' : ' + str(value))
+            else:
+                a.append('  <invalid item>')
+        a = sorted(a)
+        a.insert(0, 'SymbolTable(')
+        a[-1] += ')'
+        return '\n'.join(a)
+
+    def copy(self) -> 'SymbolTable':
+        return SymbolTable((key, node.copy())
+                           for key, node in self.items())
+
+    def serialize(self, fullname: str) -> JsonDict:
+        data = {'.class': 'SymbolTable'}  # type: JsonDict
+        for key, value in self.items():
+            # Skip __builtins__: it's a reference to the builtins
+            # module that gets added to every module by
+            # SemanticAnalyzerPass2.visit_file(), but it shouldn't be
+            # accessed by users of the module.
+            if key == '__builtins__' or value.no_serialize:
+                continue
+            data[key] = value.serialize(fullname, key)
+        return data
+
+    @classmethod
+    def deserialize(cls, data: JsonDict) -> 'SymbolTable':
+        from mypy.nodes import SymbolTableNode
+
+        assert data['.class'] == 'SymbolTable'
+        st = SymbolTable()
+        for key, value in data.items():
+            if key != '.class':
+                st[key] = SymbolTableNode.deserialize(value)
+        return st
